Split codegen into two phases: Parsing the input and generating output
So far, the parser has been pretty much single-pass: Read the input, then write the output.

That is going to be limiting going forward, so this patch splits the parsing out into parse_field, which
returns FieldDefinition. FieldDefinition has all the same fields as what we needed so far, but conveniently bundled
together.