Fix signed field setters
This is a pretty severe bug: A signed field (e.g. i8) would, if negative, set all bits above to 1 when being written to.

This patch fixes this first converting to the unsigned equivalent, then to the base data type.

So for example to write an i8 inside a u32:

value as u8 as u32